requirements.txt
# Core Streamlit and Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.6

# Data Processing and Visualization
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AI and Machine Learning
google-generativeai==0.3.2
Pillow==10.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Date and Time Utilities
python-dateutil==2.8.2

# File Processing
openpyxl==3.1.2

# Development and Testing
pytest==7.4.3
black==23.11.0

# Optional: Google Calendar Integration
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0